requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
antlr-python-runtime=4.7.2=py37hf985489_1002
aospy=0.3.1=py37hf985489_1001
appnope=0.1.2=py37hf985489_1
argon2-cffi=20.1.0=py37h271585c_2
artools=0.0.0=pypi_0
asciitree=0.3.3=py_2
async_generator=1.10=py_0
attrs=21.2.0=pyhd8ed1ab_0
autopep8=1.5.7=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
basemap=1.2.2=py37h7d89ad1_3
basemap-data-hires=1.2.2=0
beautifulsoup4=4.9.3=pyhb0f4dca_0
bleach=3.3.0=pyh44b312d_0
blosc=1.21.0=he49afe7_0
bokeh=2.4.1=pypi_0
brotlipy=0.7.0=py37h271585c_1001
bzip2=1.0.8=h0d85af4_4
c-ares=1.17.1=h0d85af4_1
ca-certificates=2021.10.8=h033912b_0
cached-property=1.5.2=pypi_0
cartopy=0.18.0=py37hec0b5f8_8
cdo=1.9.10=h28f6989_1
certifi=2021.10.8=py37hf985489_1
cf-units=3.0.1=py37h238668a_0
cf_units=2.0.1=py37h917ab60_1002
cffi=1.14.6=py37hda351d9_0
cftime=1.5.0=py37h238668a_0
chardet=4.0.0=py37hf985489_1
click=8.0.1=py37hf985489_0
cloudpickle=1.6.0=py_0
cmocean=2.0=py_3
colorcet=2.0.6=pypi_0
colorspacious=1.1.2=pyh24bf2e0_0
conda=4.10.3=py37hf985489_0
conda-package-handling=1.7.3=py37h271585c_0
cryptography=3.4.7=py37hce4a858_0
curl=7.77.0=hb861fe1_0
cycler=0.10.0=py_2
cytoolz=0.11.0=py37h271585c_3
dask=2021.7.0=pyhd8ed1ab_0
dask-core=2021.7.0=pyhd8ed1ab_0
dbus=1.13.6=ha13b53f_2
debugpy=1.3.0=py37hd8d24ac_0
decorator=5.0.9=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
distributed=2021.7.0=py37hf985489_0
docopt=0.6.2=py_1
eccodes=2.22.1=hc49948d_0
entrypoints=0.3=pyhd8ed1ab_1003
eofs=1.4.0=py_0
esmf=8.1.1=nompi_h3060d44_0
et_xmlfile=1.0.1=py_1001
expat=2.4.1=he49afe7_0
fasteners=0.16=pyhd8ed1ab_0
fftw=3.3.9=nompi_h02cd531_101
fortran-magic=0.7=py_1001
freetype=2.10.4=h4cff582_1
fsspec=2021.6.1=pyhd8ed1ab_0
geos=3.8.1=h4a8c4bd_0
geoviews=1.9.2=pypi_0
gettext=0.19.8.1=h7937167_1005
git=2.30.2=pl5320h2551975_1
git-lfs=2.13.3=h694c41f_0
glib=2.68.3=he49afe7_0
glib-tools=2.68.3=he49afe7_0
gsl=2.6=h71c5fe9_2
gsw=3.4.0=py37h9f9e6dc_1
h5py=3.4.0=pypi_0
haversine=2.3.1=pyhd8ed1ab_0
hdf4=4.2.15=hefd3b78_3
hdf5=1.10.6=nompi_hc5d9132_1114
heapdict=1.0.1=py_0
holoviews=1.14.6=pypi_0
hvplot=0.7.3=pypi_0
icu=68.1=h74dc148_0
idna=2.10=pyh9f0ad1d_0
importlib-metadata=3.10.1=py37hf985489_0
ipykernel=6.0.1=py37h85f7c60_0
ipython=7.25.0=py37h85f7c60_1
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
iris=2.4.0=py37hf985489_1
iris-sample-data=2.3.0=pyh9f0ad1d_0
jasper=1.900.1=h636a363_1006
jbig=2.1=h0d85af4_2003
jdcal=1.4.1=py_0
jedi=0.18.0=py37hf985489_2
jinja2=3.0.1=pyhd8ed1ab_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=hbcb3906_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter=1.0.0=py37hf985489_6
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_console=6.4.0=pyhd8ed1ab_0
jupyter_core=4.7.1=py37hf985489_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.1=py37h737db71_1
krb5=1.19.1=hcfbf3a7_0
lcms2=2.12=h577c468_0
lerc=2.2.1=h046ec9c_0
libaec=1.0.5=he49afe7_0
libblas=3.9.0=9_openblas
libcblas=3.9.0=9_openblas
libclang=11.1.0=default_he082bbe_1
libcurl=7.77.0=hf45b732_0
libcxx=11.1.0=habf9029_0
libdeflate=1.7=h35c211d_5
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.3=h046ec9c_2
libgfortran=5.0.0=9_3_0_h6c81a4c_22
libgfortran5=9.3.0=h6c81a4c_22
libglib=2.68.3=hd556434_0
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=9_openblas
libllvm11=11.1.0=hd011deb_2
libnetcdf=4.8.0=nompi_hb4d10b0_103
libnghttp2=1.43.0=h07e645a_0
libopenblas=0.3.15=openmp_h5e1b9a4_1
libpng=1.6.37=h7cec526_2
libpq=13.3=hea3049e_0
libsodium=1.0.18=hbcb3906_1
libssh2=1.9.0=h52ee1ee_6
libtiff=4.3.0=h1167814_1
libwebp-base=1.2.0=h0d85af4_2
libxml2=2.9.12=h93ec3fd_0
libzip=1.8.0=h8b0c345_0
line_profiler=3.3.0=py37h737db71_0
llvm-openmp=11.1.0=hda6cdc1_1
llvmlite=0.36.0=pypi_0
locket=0.2.1=pypi_0
lz4-c=1.9.3=h046ec9c_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py37h271585c_0
mat73=0.52=pypi_0
matplotlib=3.4.2=py37hf985489_0
matplotlib-base=3.4.2=py37hb018525_0
matplotlib-inline=0.1.2=pyhd8ed1ab_2
mistune=0.8.4=py37h271585c_1004
mock=4.0.3=py37hf985489_1
monotonic=1.5=py_0
msgpack-python=1.0.2=py37h737db71_1
mysql-common=8.0.25=h694c41f_2
mysql-libs=8.0.25=h115446f_2
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.1.0=py37hf985489_0
nbformat=5.1.3=pyhd8ed1ab_0
nco=5.0.0=h2e37db4_0
ncurses=6.2=h2e338ed_4
nest-asyncio=1.5.1=pyhd8ed1ab_0
netcdf-fortran=4.5.3=nompi_hff36437_104
netcdf4=1.5.7=nompi_py37h75b5d8a_100
notebook=6.4.0=pyha770c72_0
nspr=4.30=hcd9eead_0
nss=3.67=h31e2bf1_0
numba=0.53.1=pypi_0
numcodecs=0.8.0=py37hd8d24ac_0
numexpr=2.7.3=py37hff173d8_0
numpy=1.21.0=py37h84c02c4_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.4.0=h6e7aa92_1
openpyxl=3.0.7=pyhd8ed1ab_0
openssl=1.1.1l=h0d85af4_0
ossuuid=1.6.2=h0a44026_1000
packaging=21.0=pyhd8ed1ab_0
pandas=1.2.4=py37h23ab428_0
pandoc=2.14.0.3=h0d85af4_0
pandocfilters=1.4.2=py_1
panel=0.12.4=pypi_0
param=1.11.1=pypi_0
parso=0.8.2=pyhd8ed1ab_0
partd=1.2.0=pyhd8ed1ab_0
patsy=0.5.1=py_0
pcre=8.45=he49afe7_0
perl=5.32.1=0_h0d85af4_perl5
pexpect=4.8.0=pyh9f0ad1d_2
pickle5=0.0.11=py37h271585c_1
pickleshare=0.7.5=py_1003
pillow=8.3.1=py37h76dc067_0
pip=21.1.3=pyhd8ed1ab_0
proj=7.2.0=h78d1473_2
prometheus_client=0.11.0=pyhd8ed1ab_0
prompt-toolkit=3.0.19=pyha770c72_0
prompt_toolkit=3.0.19=hd8ed1ab_0
psutil=5.8.0=py37h271585c_1
ptyprocess=0.7.0=pyhd3deb0d_0
pycodestyle=2.7.0=pyhd8ed1ab_0
pycosat=0.6.3=py37h271585c_1006
pycparser=2.20=pyh9f0ad1d_2
pyct=0.4.8=pypi_0
pydap=3.2.2=pyh9f0ad1d_1001
pygments=2.9.0=pyhd8ed1ab_0
pyke=1.1.1=pyhd8ed1ab_1004
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyproj=3.1.0=py37hf99dfa2_3
pyqt=5.12.3=py37hf985489_7
pyqt-impl=5.12.3=py37h5ae2165_7
pyqt5-sip=4.19.18=py37h5ab9d7d_7
pyqtchart=5.12=py37h5ae2165_7
pyqtwebengine=5.12.1=py37h5ae2165_7
pyrsistent=0.17.3=py37h271585c_2
pyshp=2.1.3=pyh44b312d_0
pysocks=1.7.1=py37hf985489_3
pytables=3.6.1=py37h6293439_3
python=3.7.10=h7728216_100_cpython
python-cdo=1.5.4=pyhd8ed1ab_0
python-dateutil=2.8.1=py_0
python-xxhash=2.0.2=py37h271585c_0
python_abi=3.7=2_cp37m
pytz=2021.1=pyhd8ed1ab_0
pyviz-comms=2.1.0=pypi_0
pyyaml=5.4.1=py37hf967b71_0
pyzmq=22.1.0=py37h8f778e5_0
qt=5.12.9=h126340a_4
qtconsole=5.1.1=pyhd8ed1ab_0
qtpy=1.9.0=py_0
readline=8.1=h05e3726_0
requests=2.25.1=pyhd3deb0d_0
ruamel_yaml=0.15.80=py37h271585c_1004
scikit-learn=0.24.2=py37h11db1d0_0
scipy=1.7.0=py37h866764c_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
seawater=3.3.4=py_1
send2trash=1.7.1=pyhd8ed1ab_0
setuptools=49.6.0=py37hf985489_3
shapely=1.7.1=py37h2a6e0eb_1
six=1.16.0=pyh6c4a22f_0
sortedcontainers=2.4.0=pyhd8ed1ab_0
soupsieve=2.0.1=py_1
sqlite=3.36.0=h23a322b_0
statsmodels=0.12.2=py37h183f225_0
tblib=1.7.0=pyhd8ed1ab_0
tempest-remap=2.1.0=hf899ff6_0
terminado=0.10.1=py37hf985489_0
testpath=0.5.0=pyhd8ed1ab_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=h0419947_1
toml=0.10.2=pyhd8ed1ab_0
toolz=0.11.1=py_0
tornado=6.1=py37h271585c_1
tqdm=4.61.2=pyhd8ed1ab_1
traitlets=5.0.5=py_0
typing_extensions=3.10.0.0=pyha770c72_0
udunits2=2.2.27.27=h664449b_1
urllib3=1.26.6=pyhd8ed1ab_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
webob=1.8.7=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
widgetsnbextension=3.5.1=py37hf985489_4
xarray=0.18.2=pyhd8ed1ab_0
xlrd=2.0.1=pyhd8ed1ab_3
xscale=0.1.dev0+35.ga804866=pypi_0
xxhash=0.8.0=h35c211d_3
xz=5.2.5=haf1e3a3_1
yaml=0.2.5=haf1e3a3_0
zarr=2.8.3=pyhd8ed1ab_0
zeromq=4.3.4=h1c7c35f_0
zict=2.0.0=py_0
zipp=3.5.0=pyhd8ed1ab_0
zlib=1.2.11=h7795811_1010
zstd=1.5.0=h582d3a0_0
